# src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "StockStream API",
    "version": "1.0.0",
    "description": "Real-time stock data API supporting VN (Yahoo Finance), US stocks and crypto (Finnhub). Provides OHLCV candles with SQLite caching, live quotes via WebSocket, symbol search, and latency analytics."
  },
  "servers": [
    { "url": "http://localhost:3001", "description": "Local development" }
  ],
...
  "paths": {
...
    "/api/symbols": {
      "get": {
        "tags": ["symbols"],
        "summary": "Full symbol list",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "schema": { "type": "string", "enum": ["vn","us","crypto"] },
            "description": "Filter by market type"
          }
        ],
        "responses": {
          "200": {
            "description": "All available symbols",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/SymbolInfo" }
                }
              }
            }
          }
        }
      }
    },
    "/api/symbols/search": {
      "get": {
        "tags": ["symbols"],
        "summary": "Search symbols by name or ticker",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Case-insensitive search against symbol and name",
            "example": "vnm"
          },
          {
            "name": "type",
            "in": "query",
            "schema": { "type": "string", "enum": ["vn","us","crypto"] },
            "description": "Optional market filter"
          }
        ],
        "responses": {
          "200": {
            "description": "Up to 20 matching symbols",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "maxItems": 20,
                  "items": { "$ref": "#/components/schemas/SymbolInfo" }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SymbolInfo": {
        "type": "object",
        "required": ["symbol","name","exchange","type"],
        "properties": {
          "symbol":   { "type": "string", "example": "VN:VNM" },
          "name":     { "type": "string", "example": "Vinamilk" },
          "exchange": { "type": "string", "example": "HOSE" },
          "type":     { "type": "string", "enum": ["vn","us","crypto"] }
        }
      },
...
    },
...
  },
...
}